+++ src/yapss/_private/mseipopt/__init__.py
@@ -9,5 +9,6 @@
 """The most simple ever IPOPT interface."""
 
 from .bare import load_library, use_library
+from .library import initialize_ipopt
 
-__all__ = ["load_library", "use_library"]
+__all__ = ["initialize_ipopt", "load_library", "use_library"]
